First version
Tested to generate object 3 and 3417 downloaded from the OMA Registry.
Have not done any runtime testing as of now. Although the generated code
anyway need some attention before being ready to use so I do not think
it is that important.

It is tested to compile without any warnings and produces code very
close to the already implemented instance of object 3. That is the usage
details which are not described in the XML is not generated but that
would not be possible as of now anyway.